+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# TBA
+
+- Add `BitsyClient.truncate` to trigger truncation on Bitsy (see [this](https://github.com/ramontayag/bitsy/issues/6))
+
 # v0.1.0
 
 - Add `BitsyClient::Sync` to trigger a transaction sync manually